% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_eucalipto}
\alias{ad_eucalipto}
\title{Example of Eucalyptus Plant Data}
\usage{
ad_eucalipto()
}
\value{
A data frame
}
\description{
The data contains measurements of diameter at breast height (DBH)
using both a caliper (dap_suta) and a tape measure (dap_trena), as well
as the circumference of eucalyptus plants. A total of 70 eucalyptus
plants were evaluated, distributed across 5 rows with 14 plants per row.
}
