% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_linhaca}
\alias{ad_linhaca}
\title{Example of Flax Plant Data}
\usage{
ad_linhaca()
}
\value{
A data frame
}
\description{
The data contains evaluations of flax plants sown at four different times
(E1 to E4). The experiment involved sampling between 51 and 66 plants
per sowing time. Various plant traits were analyzed, including:
}
\details{
\itemize{
\item ac: capsule height
\item ap: plant height
\item nc: number of capsules
\item ng: number of grains
\item areac: capsule area
\item compg: grain length
\item largg: grain width
\item nr: number of branches
\item mc: capsule mass
\item rgpla: grain yield per plant
\item icc: capsule harvest index
\item ngcap: number of grains per capsule
\item mmg: thousand-grain mass
}
}
