% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_normal}
\alias{ad_normal}
\title{Example of Normal data}
\usage{
ad_normal()
}
\value{
A data frame
}
\description{
The data contains the length of 205.043 flax grains that follows a normal
distribution
}
