% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_quali}
\alias{ad_quali}
\title{Example of Qualitative data}
\usage{
ad_quali()
}
\value{
A data frame
}
\description{
The data contains the area (AF) and dry matter (MST) of chicory plants
evaluated at different levels of solar radiation (50\%, 70\%, and 100\%).
The experiment was conducted in a randomized complete block design with
four replications.
}
