% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_quali_quali_ci}
\alias{ad_quali_quali_ci}
\title{Example of Qualitative x Qualitative data (with significant interaction)}
\usage{
ad_quali_quali_ci()
}
\value{
A data frame
}
\description{
The data contains two qualitative factors: sulfur (2 levels) and nitrogen
splitting (3 levels), with significant interaction.
}
