% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_quali_quali_si}
\alias{ad_quali_quali_si}
\title{Example of Qualitative x Qualitative data (no significant interaction)}
\usage{
ad_quali_quali_si()
}
\value{
A data frame
}
\description{
The data contains two qualitative factors: evaluation days (3 levels)
and solar radiation (3 levels), with no significant interaction.
}
