% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_quali_quanti_ci}
\alias{ad_quali_quanti_ci}
\title{Example of Qualitative x Quantitative data (with significant interaction)}
\usage{
ad_quali_quanti_ci()
}
\value{
A data frame
}
\description{
The data contains two factors: hybrids (2 levels, qualitative) and nitrogen
doses (5 levels, quantitative), with significant interaction.
}
