% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_quali_quanti_si}
\alias{ad_quali_quanti_si}
\title{Example of Qualitative x Quantitative data (no significant interaction)}
\usage{
ad_quali_quanti_si()
}
\value{
A data frame
}
\description{
The data contains two factors: hybrids (2 levels, qualitative) and nitrogen
doses (5 levels, quantitative), with no significant interaction.
}
