% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_quanti_1}
\alias{ad_quanti_1}
\title{Example of Quantitative data (linear trend)}
\usage{
ad_quanti_1()
}
\value{
A data frame
}
\description{
The data contains maize yield evaluated at different nitrogen levels,
showing a linear trend.
}
