% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ad_quanti_2}
\alias{ad_quanti_2}
\title{Example of Quantitative data (quadratic trend)}
\usage{
ad_quanti_2()
}
\value{
A data frame
}
\description{
The data contains oat yield evaluated at different nitrogen levels,
showing a quadratic trend.
}
